+++ src/com.unity.process-server/CHANGELOG.md
@@ -1,3 +1,14 @@
 # Changelog
 
 ## [VERSION] - DATE
+
+- Add ProcessOptions to RpcServerTask
+
+## [0.1.27-preview] - 2019-12-17
+
+- ILRepack StreamRPC DLLs into one bundle
+- Fixes
+
+## [0.1.0-preview] - 2019-12-12
+
+- com.unity.process-server